Play struct tetris, win bytes.
git-svn-id: https://dev.csoft.net/agar/trunk@10350 86a84d31-9702-0410-8f1e-a60cab7bd424

+++ vg/vg_arc.h
@@ -3,7 +3,7 @@
 typedef struct vg_arc {
 	struct vg_node _inherit;
 	VG_Point *_Nullable p;		/* Centerpoint */
-	float r;			/* Arc radius */
+	double r;			/* Arc radius */
 	float a1;			/* Start angle (degs) */
 	float a2;			/* End angle (degs) */
 } VG_Arc;
@@ -17,5 +17,5 @@ VG_Arc *_Nonnull VG_ArcNew(void *_Nullable, VG_Point *_Nonnull,
                            float, float,float);
 
 void VG_ArcCenter(VG_Arc *_Nonnull, VG_Point *_Nonnull);
-void VG_ArcRadius(VG_Arc *_Nonnull, float);
+void VG_ArcRadius(VG_Arc *_Nonnull, double);
 __END_DECLS
